+++ Sources/SMHSSchedule (iOS)/Views/TodayView/TodayViewViewModel.swift
@@ -10,5 +10,6 @@ import Foundation
 
 class TodayViewViewModel: ObservableObject {
     @Published var showEditModal = false
-    
+    @Published var selectionMode: PeriodCategory = .firstLunch
+
 }