+++ Sources/Shared/Styles/HighlightButtonStyle.swift
@@ -8,13 +8,28 @@
 import SwiftUI
 
 struct HighlightButtonStyle: ButtonStyle {
+    var foregroundColor: Color = .white
+    var backgroundColor: Color = .primary
+    var horizontalPadding: CGFloat = 100
+    var autoWidth: CGFloat {
+        min(CGFloat(400), UIScreen.screenWidth - horizontalPadding)
+    }
     func makeBody(configuration: Configuration) -> some View {
         configuration.label
-            .frame(width: min(400, UIScreen.screenWidth - CGFloat(100)))
+            .frame(width: autoWidth)
             .padding()
-            .background(Color.primary)
-            .foregroundColor(Color.white)
+            .background(backgroundColor) 
+            .foregroundColor(foregroundColor)
             .clipShape(Capsule(style: .continuous))
+            .animatedEffects(configuration)
+    }
+    
+    
+}
+
+fileprivate extension View {
+    func animatedEffects(_ configuration: ButtonStyleConfiguration) -> some View {
+        self
             .brightness(configuration.isPressed ? Double(-0.1) : Double(0))
             .scaleEffect(configuration.isPressed ? CGFloat(0.97) : CGFloat(1))
             .animation(Animation.easeInOut)