+++ Sources/Shared/Views/SearchView/SearchView.swift
@@ -21,7 +21,7 @@ struct SearchView: View {
             ZStack {
                 if searchText.isEmpty {
                     InformationCardsView()
-                        
+                        .aboutFooter(showDivider: false)
                 }
                 else {
                     SearchResultView(searchText: $searchText,