+++ Sources/Shared/Views/SettingsView/AboutDevView.swift
@@ -1,0 +1,20 @@
+//
+//  AboutDevView.swift
+//  SMHSSchedule
+//
+//  Created by Jevon Mao on 5/24/21.
+//
+
+import SwiftUI
+
+struct AboutDevView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+struct AboutDevView_Previews: PreviewProvider {
+    static var previews: some View {
+        AboutDevView()
+    }
+}